# openapi/studio.yaml
openapi: 3.0.3
info:
  title: API Develop Studio
  version: 1.0.1
  description: Studio management API. Existing business validation and error contracts
    are retained.
paths:
  /api/projects:
    get:
      operationId: list_projects
      tags:
      - projects
      summary: list projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReferenceList'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.list_documents
      x-studio-parts: '[''projects'']'
      x-studio-route: /api/projects
      x-studio-upload: false
      x-studio-alias: true
      parameters:
      - name: project
        in: query
        required: false
        schema: &id001
          type: string
  /api/projects/{reference}/revisions:
    get:
      operationId: list_projects_revisions
      tags:
      - projects
      summary: list projects revisions
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RevisionList'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.list_revisions
      x-studio-parts: '[''projects'']'
      x-studio-route: /api/projects/{reference:path}/revisions
      x-studio-upload: false
      x-studio-alias: true
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
  /api/projects/{reference}:
    get:
      operationId: get_projects
      tags:
      - projects
      summary: get projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectView'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.get_document
      x-studio-parts: '[''projects'']'
      x-studio-route: /api/projects/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
    put:
      operationId: save_projects
      tags:
      - projects
      summary: save projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.save_document
      x-studio-parts: '[''projects'']'
      x-studio-route: /api/projects/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
    delete:
      operationId: delete_projects
      tags:
      - projects
      summary: delete projects
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeleteResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Policy denied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Revision conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Storage unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-studio-handler: documents.delete_document
      x-studio-parts: '[''projects'']'
      x-studio-route: /api/projects/{reference:path}
      x-studio-upload: false
      x-studio-alias: false
      parameters:
      - name: reference
        in: path
        required: true
        schema: *id001
        description: Slash-separated document or attachment reference.
...
components:
  schemas:
    ErrorResponse:
      type: object
      properties:
        error: *id001
        code: *id001
        currentRevision: &id002
          type: integer
        runId: *id001
      required:
      - error
    StorageMetadata:
      type: object
      properties:
        id:
          type: string
        revision: *id002
        created_by: *id001
        updated_by: *id001
        created_at: *id001
        updated_at: *id001
        content_hash: *id001
        projectionPending:
          type: boolean
    SecretInput:
      type: object
      properties:
        value:
          type: string
          writeOnly: true
        preserve: &id003
          type: boolean
    SecretState:
      type: object
      properties:
        configured: *id003
        value:
          type: string
          description: Only the explicitly public bundled Example fixture.
      required:
      - configured
    VariablesInput:
      type: object
      properties:
        plain:
          type: object
          additionalProperties: *id001
        secret:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SecretInput'
    VariablesView:
      type: object
      properties:
        plain:
          type: object
          additionalProperties: *id001
        secret:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/SecretState'
    AuthProfile:
      type: object
      properties:
        type:
          type: string
          enum:
          - No Auth
          - API Key
          - Basic Auth
          - Bearer Token
        key: *id001
        value: *id001
        addTo: *id001
        username: *id001
        password: *id001
        token: *id001
      required:
      - type
    EnvironmentInput:
      type: object
      properties:
        base_url: *id001
        variables:
          $ref: '#/components/schemas/VariablesInput'
        auth_profiles:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AuthProfile'
      required:
      - base_url
    EnvironmentView:
      type: object
      properties:
        base_url: *id001
        variables:
          $ref: '#/components/schemas/VariablesView'
        auth_profiles:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AuthProfile'
      required:
      - base_url
    NamedBaseUrl:
      type: object
      properties:
        name: *id001
        url: *id001
      required:
      - name
      - url
    TransportSettings:
      type: object
      properties:
        use_proxy: *id003
        verify: *id003
        proxy: *id001
        http_proxy: *id001
        https_proxy: *id001
    DocumentBundle:
      type: object
      properties:
        entrypoint: *id001
        files:
          type: object
          additionalProperties: *id001
      required:
      - entrypoint
      - files
    DocumentFile:
      type: object
      properties:
        name: *id001
        document: &id010
          type: object
          additionalProperties: true
      required:
      - name
      - document
    ProjectInput:
      type: object
      properties:
        name: *id001
        base_url: *id001
        base_urls: &id004
          type: array
          items:
            $ref: '#/components/schemas/NamedBaseUrl'
        docs_url: *id001
        docs_file: &id005
          $ref: '#/components/schemas/DocumentFile'
        docs_bundle: &id006
          $ref: '#/components/schemas/DocumentBundle'
        advanced: &id007
          $ref: '#/components/schemas/TransportSettings'
        default_environment: *id001
        auth_profiles: &id008
          type: object
          additionalProperties:
            $ref: '#/components/schemas/AuthProfile'
        _storage: &id009
          $ref: '#/components/schemas/StorageMetadata'
        variables:
          $ref: '#/components/schemas/VariablesInput'
        environments:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EnvironmentInput'
      required:
      - name
      - base_url
    ProjectView:
      type: object
      properties:
        name: *id001
        base_url: *id001
        base_urls: *id004
        docs_url: *id001
        docs_file: *id005
        docs_bundle: *id006
        advanced: *id007
        default_environment: *id001
        auth_profiles: *id008
        _storage: *id009
        variables:
          $ref: '#/components/schemas/VariablesView'
        environments:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/EnvironmentView'
      required:
      - name
      - base_url
...
    ReferenceList:
      type: object
      properties:
        items:
          type: array
          items: *id001
        details:
          type: object
          additionalProperties: *id010
      required:
      - items
    RevisionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StorageMetadata'
      required:
      - items
    SaveResponse:
      type: object
      properties:
        path: *id001
        _storage:
          $ref: '#/components/schemas/StorageMetadata'
      required:
      - path
      - _storage
    DeleteResponse:
      type: object
      properties:
        deleted: *id001
        deleted_pipelines:
          type: array
          items: *id001
      required:
      - deleted
...